Move DisplaySprites into non-matching

+++ include/gba/defines.h
@@ -62,7 +62,7 @@
 #define TILE_SIZE_8BPP 64
 
 // NOTE/TODO: Maybe this should somewhere else?
-#define GET_TILE_NUM_COMMON(vramPtr, tileSize) (((size_t)(vramPtr) - OBJ_VRAM0) / (tileSize))
+#define GET_TILE_NUM_COMMON(vramPtr, tileSize) ((((uintptr_t)(vramPtr)) - (uintptr_t)OBJ_VRAM0) / (tileSize))
 #define GET_TILE_NUM(vramPtr) GET_TILE_NUM_COMMON((vramPtr), TILE_SIZE_4BPP) 
 
 #define TOTAL_OBJ_TILE_COUNT 1024
